+++ cfdm/read_write/netcdf/netcdfwrite.py
@@ -6222,6 +6222,13 @@ class NetCDFWrite(IOWrite):
                             authority="",
                             path=uri.path,
                         )
+                        fragment = uri.fragment
+                        if fragment is not None:
+                            # Append a URI fragment. Do this with a
+                            # string-append, rather than via
+                            # `uricompose` in case the fragment
+                            # contains more than one # character.
+                            aggregation_file_directory += f"#{fragment}"
 
                     g["aggregation_file_directory"] = (
                         aggregation_file_directory
@@ -6281,8 +6288,15 @@ class NetCDFWrite(IOWrite):
                     filename = uricompose(
                         scheme="file",
                         authority="",
-                        path=abspath(uri.path),
+                        path=uri.path,
                     )
+                    fragment = uri.fragment
+                    if fragment is not None:
+                        # Append a URI fragment. Do this with a
+                        # string-append, rather than via `uricompose`
+                        # in case the fragment contains more than one
+                        # # character.
+                        filename += f"#{fragment}"
 
                 if uri_relative:
                     scheme = uri.scheme